OLH-1036: Remove hello world Lambda infrastructure
We had this because we needed something in the template so that the
pipeline would run. We're about to add some new AWS resources so we
don't need this example anymore.

I'm choosing to leave the lambda code in place so we have a template to
copy when we put a real lambda in this repository. We should remove the
example lambda at that point