OLH-1036: Create a backup wrapper key
We'll be using this key from our main AWS account to envelope encrypt
user data. Having the backup key in a separate AWS account gives us some
resilience in a disaster recovery situation.

This commit doesn't set any permissions for the other AWS account to use
this key; we'll do that in future commits.