BAU: Delete the example lambda
We initially included this lambda as an example for how we'd set this
repo up. We've ended up not using it as much as we initially thought,
and if we did add lambdas we'd use a different structure using a single
NPM package anyway.

Additionally, there's a critical security alert out for babel which we
use here[^1]. This lambda's not deployed anywhere so we're not
vulnerable, but it's on our alert dashboard and we should clean it up.

The tidiest solution to unused code is to delete it. We have the git
history if we ever want to see what the example lambda did, the github
actions etc.

[^1]: https://github.com/govuk-one-login/di-account-data-backend/security/dependabot/6