BAU: Remove Node install steps from Actions
We removed all the Lambda code from this repo in #16 - it wasn't used
and we were getting security warnings.
In that we missed that the actions also install Node and rely on a now
deleted file. Thankfully we can just also delete the affected steps from
the action definitions as well.